Removed unnecessary services; Added interfaces implementations

+++ ITestApp/ITestApp.Data.Models/Test.cs
@@ -18,7 +18,7 @@ namespace ITestApp.Data.Models
         [Required]
         public string Name { get; set; }
 
-        public double RequiredTime { get; set; }
+        public int RequiredTime { get; set; }
 
         public string AuthorId { get; set; }
         public User Author { get; set; }
